# Conflicts:
#	src/Http/Controllers/CP/Assets/AssetContainersController.php

+++ src/Http/Controllers/CP/Assets/AssetContainersController.php
@@ -25,7 +25,9 @@ class AssetContainersController extends CpController
                 'allow_uploads' => $container->allowUploads(),
                 'create_folders' => $container->createFolders(),
                 'edit_url' => $container->editUrl(),
-                'delete_url' => $container->deleteUrl()
+                'delete_url' => $container->deleteUrl(),
+                'can_edit' => User::current()->can('edit', $container),
+                'can_delete' => User::current()->can('delete', $container),
             ];
         })->values();
 
@@ -133,13 +135,13 @@ class AssetContainersController extends CpController
 
     public function destroy($container)
     {
-        // TODO: auth
+        $this->authorize('delete', $container, 'You are not authorized to delete asset containers.');
 
         $container->delete();
 
         return [
             'message' => 'Container deleted',
-            'redirect' => cp_route('asset-containers.index')
+            'redirect' => cp_route('assets.index')
         ];
     }
